Set up an add to playlist button component
Set up an add to playlist button component. Am able to get
an object with user_id and episode_id in a "stop-gap" way,
but this is not the correct way to do it and I need to
figure out how to set it up to work properly with my back-end,
as the relationships are awkward.